Use androguard new permissions system
See https://github.com/androguard/androguard/commit/828127780111080a2dc78cb13578ede797cb3575
for more information.

Androguard shifted to a new handling of permissions, allowing them to
have different API level (24 for Android 7.0) which I chose as a
default. It is just a matter of changing the constant if you consider
anoyther version should be the default.