Use {% url %} blocks and reverse.
It is more generic and advised
( see https://docs.djangoproject.com/en/1.11/topics/http/urls/#reverse-resolution-of-urls )
to url {% url %} blocks in templates and reverse in python code.

This is more stable, allows to run the application on a suburl (like
$TLD/exodus/ ). Moreover, this allows to change all urls only by
changing the urls.py file, without having to go over all the code to
find any instance of urls.